pyproject.toml
[project]
name = "obsidian-knowledge"
version = "0.1.0"
description = "Agentic knowledge server backed by Obsidian and Elasticsearch"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "fastmcp>=2.0",
    "elasticsearch>=8.17",
    "python-frontmatter>=1.1",
    "anthropic>=0.52",
    "watchfiles>=1.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["backend"]
testpaths = ["tests"]
